100-prime_factor.c

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -8,10 +8,10 @@
 
 int main(void)
 {
-		long long n = 612852475143;
-		long long largestPrime = 0;
+		long n = 612852475143;
+		long largestPrime = 0;
 
-		for (long long i = 2; i <= n;)
+		for (long i = 2; i <= n;)
 		{
 			if (n % i == 0)
 			{
@@ -23,6 +23,6 @@ int main(void)
 				i++;
 			}
 		}
-	printf("%lld\n", largestPrime);
+	printf("%ld\n", largestPrime);
 	return (0);
 }
